Make Obbjects Data Contracts. Extract Interface out of Node and Binary Tree. Finish Tests. Implement Cloning. Clean Up Test project structure
