+++ ScaffelPikeDerivatives/Interfaces/IBinaryTree.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace ScaffelPikeDerivatives.Objects
+namespace ScaffelPikeDerivatives.Objects.Interfaces
 {
-  public interface IBinaryTree<T> where T : IComparable<T>
+  public interface IBinaryTree<T> : ICloneable where T : IComparable<T>
   {
     int Count { get; }
     int Time  { get; }
