+++ ScaffelPikeDerivatives/ScaffelPikeDerivatives.csproj
@@ -33,6 +33,7 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -42,7 +43,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Objects\BinaryTree.cs" />
-    <Compile Include="IBinaryTree.cs" />
+    <Compile Include="Interfaces\IBinaryTree.cs" />
+    <Compile Include="Interfaces\INode.cs" />
     <Compile Include="Objects\Node.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
