+++ ScaffelPikeTests/Derivatives/Objects/BinaryTreeTests/BinaryTreeTest.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using ScaffelPikeTests.Derivatives.Objects;
+using Xunit;
+
+namespace BinaryTreeTests
+{
+  public class BuildingAndAccess
+  {
+    [Fact]
+    public void InsertAndGet()
+    {
+      //arrange
+      var bt = BinaryTreeTestFactory.GenerateTimeTwoTree();
+
+      //act and assert
+      Assert.Equal("root",bt.GetAt(new bool[] { }).Data);
+      Assert.Equal("h", bt.GetAt(new bool[] { true }).Data);
+      Assert.Equal("t", bt.GetAt(new bool[] { false }).Data);
+      Assert.Equal("hh", bt.GetAt(new bool[] { true, true }).Data);
+      Assert.Equal("ht", bt.GetAt(new bool[] { true, false }).Data);
+      Assert.Equal("tt", bt.GetAt(new bool[] { false, false }).Data);
+      Assert.Equal("th", bt.GetAt(new bool[] { false, true }).Data);
+      Assert.Equal(7, bt.Count);
+      Assert.Equal(2, bt.Time);
+    }
+    [Fact]
+    public void GetAnItemThatDoesntExist()
+    {
+      //arange
+      var nonExistentPath = new bool[] { false, false, false };
+      var bt = BinaryTreeTestFactory.GenerateTimeTwoTree();
+
+      //act and assert
+      var ex = Assert.Throws<ArgumentException>(() => bt.GetAt(nonExistentPath));
+      Assert.Equal(ex.ParamName, "path");
+    }
+  }
+}
